requirements.txt
celery==5.2.7
redis==4.5.4
python-dotenv==1.0.0
requests==2.32.0
